@@ -9,8 +9,9 @@
     "google-map-react": "^2.1.10",
     "qs": "^6.10.3",
     "rc-slider": "^9.7.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-content-loader": "^6.2.0",
+    "react-dom": "^18.0.0",
     "react-query": "^3.34.16",
     "react-router-dom": "^6.2.1"
   },
