@@ -1,5 +1,4 @@
-import React from "react";
+import Dropdown from "./Dropdown";
 
 export default function FacetsPreloader(props) {
-
 }
