@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import Map from "../components/Map";
 import NavBar from "../components/NavBar";
 
